CARL Persistence -> Remove resources before inserting into the database to avoid issues while updating.
When updating a Resource in such a way that an element is removed from one of its list fields,
we did not remove these elements from the database.

This solution is bad performance-wise, but much easier to implement correctly.
A proper solution would only delete the list elements which actually got removed.